tests/wpt/web-platform-tests/css/css-break/parsing/box-decoration-break-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: getComputedValue().boxDecorationBreak</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-box-decoration-break">
<meta name="assert" content="box-decoration-break computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("box-decoration-break", "slice");
test_computed_value("box-decoration-break", "clone");
</script>
</body>
</html>
